Adding Fisherman, InnKeeper, and Stone
I have added the New job of Fishermen, the new destination of pond, the
new resource of fish, the new job of inn keeper, the new building of the
tavern, the new resource of meal, and the new resource of stone.  I have
added all these values to the city including a second mine and second
miner for stone.